TSIS4/requirements.txt
pygame>=2.5
psycopg2-binary>=2.9

